refactor: mission1 폴더링
