+++ src/main/java/org/prgms/springbootjpa/mission1/customer/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
-package org.prgms.springbootjpa.customer.repository;
+package org.prgms.springbootjpa.mission1.customer.repository;
 
-import org.prgms.springbootjpa.customer.domain.CustomerEntity;
+import org.prgms.springbootjpa.mission1.customer.domain.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<CustomerEntity, Long> {
