+++ src/main/java/org/prgms/springbootjpa/mission1/customer/repository/CustomerXmlMapper.java
@@ -1,7 +1,7 @@
-package org.prgms.springbootjpa.customer.repository;
+package org.prgms.springbootjpa.mission1.customer.repository;
 
 import org.apache.ibatis.annotations.Mapper;
-import org.prgms.springbootjpa.customer.domain.Customer;
+import org.prgms.springbootjpa.mission1.customer.domain.Customer;
 
 import java.util.List;
