+++ src/test/java/org/prgms/springbootjpa/mission1/repository/CustomerRepositoryTest.java
@@ -1,10 +1,10 @@
-package org.prgms.springbootjpa.repository;
+package org.prgms.springbootjpa.mission1.repository;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.prgms.springbootjpa.customer.repository.CustomerRepository;
-import org.prgms.springbootjpa.customer.domain.CustomerEntity;
+import org.prgms.springbootjpa.mission1.customer.repository.CustomerRepository;
+import org.prgms.springbootjpa.mission1.customer.domain.CustomerEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
